+++ sync/src/synchronizer/block_fetcher.rs
@@ -1,6 +1,6 @@
 use crate::block_status::BlockStatus;
 use crate::synchronizer::Synchronizer;
-use crate::types::{ActiveChain, HeaderView, IBDState};
+use crate::types::{ActiveChain, BlockNumberAndHash, HeaderIndex, IBDState};
 use ckb_constant::sync::{
     BLOCK_DOWNLOAD_WINDOW, CHECK_POINT_WINDOW, INIT_BLOCKS_IN_TRANSIT_PER_PEER,
 };
@@ -35,16 +35,15 @@ impl<'a> BlockFetcher<'a> {
         inflight.peer_can_fetch_count(self.peer) == 0
     }
 
-    pub fn is_better_chain(&self, header: &HeaderView) -> bool {
-        header.is_better_than(self.active_chain.total_difficulty())
-    }
-
-    pub fn peer_best_known_header(&self) -> Option<HeaderView> {
+    pub fn peer_best_known_header(&self) -> Option<HeaderIndex> {
         self.synchronizer.peers().get_best_known_header(self.peer)
     }
 
-    pub fn update_last_common_header(&self, best_known: &HeaderView) -> Option<core::HeaderView> {
-        // Bootstrap quickly by guessing a parent of our best tip is t
+    pub fn update_last_common_header(
+        &self,
+        best_known: &BlockNumberAndHash,
+    ) -> Option<BlockNumberAndHash> {
+        // Bootstrap quickly by guessing an ancestor of our best tip is forking point.
         // Guessing wrong in either direction is not a problem.
         let mut last_common =
             if let Some(header) = self.synchronizer.peers().get_last_common_header(self.peer) {
@@ -53,14 +52,14 @@ impl<'a> BlockFetcher<'a> {
                 let tip_header = self.active_chain.tip_header();
                 let guess_number = min(tip_header.number(), best_known.number());
                 let guess_hash = self.active_chain.get_block_hash(guess_number)?;
-                self.active_chain.get_block_header(&guess_hash)?
+                (guess_number, guess_hash).into()
             };
 
         // If the peer reorganized, our previous last_common_header may not be an ancestor
         // of its current tip anymore. Go back enough to fix that.
         last_common = self
             .active_chain
-            .last_common_ancestor(&last_common, best_known.inner())?;
+            .last_common_ancestor(&last_common, best_known)?;
 
         self.synchronizer
             .peers()
@@ -88,7 +87,14 @@ impl<'a> BlockFetcher<'a> {
                 // Here we need to first try search from headermap, if not, fallback to search from the db.
                 // if not search from db, it can stuck here when the headermap may have been removed just as the block was downloaded
                 if let Some(header) = self.synchronizer.shared.get_header_view(&hash, None) {
-                    state.peers().may_set_best_known_header(self.peer, header);
+                    let header_index = HeaderIndex::new(
+                        header.number(),
+                        header.hash(),
+                        header.total_difficulty().clone(),
+                    );
+                    state
+                        .peers()
+                        .may_set_best_known_header(self.peer, header_index);
                 } else {
                     state.peers().insert_unknown_header_hash(self.peer, hash);
                     break;
@@ -98,24 +104,24 @@ impl<'a> BlockFetcher<'a> {
 
         // This peer has nothing interesting.
         let best_known = self.peer_best_known_header()?;
-        if !self.is_better_chain(&best_known) {
+        if !best_known.is_better_than(self.active_chain.total_difficulty()) {
             // Advancing this peer's last_common_header is unnecessary for block-sync mechanism.
             // However, RPC `get_peers`, returns peers information which includes
             // last_common_header, is expected to provide a more realistic picture. Hence here we
             // specially advance this peer's last_common_header at the case of both us on the same
             // active chain.
             if self.active_chain.is_main_chain(&best_known.hash()) {
-                let last_common = best_known;
                 self.synchronizer
                     .peers()
-                    .set_last_common_header(self.peer, last_common.into_inner());
+                    .set_last_common_header(self.peer, best_known.number_and_hash());
             }
 
             return None;
         }
 
+        let best_known = best_known.number_and_hash();
         let last_common = self.update_last_common_header(&best_known)?;
-        if &last_common == best_known.inner() {
+        if last_common == best_known {
             return None;
         }
 
@@ -149,7 +155,7 @@ impl<'a> BlockFetcher<'a> {
                     // So we can skip the search of this space directly
                     self.synchronizer
                         .peers()
-                        .set_last_common_header(self.peer, header.clone());
+                        .set_last_common_header(self.peer, (&header).into());
                     end = min(best_known.number(), header.number() + BLOCK_DOWNLOAD_WINDOW);
                     break;
                 } else if status.contains(BlockStatus::BLOCK_RECEIVED) {