+++ sync/src/synchronizer/block_process.rs
@@ -1,7 +1,4 @@
-use crate::{
-    block_status::BlockStatus, synchronizer::Synchronizer, utils::is_internal_db_error, Status,
-    StatusCode,
-};
+use crate::{synchronizer::Synchronizer, utils::is_internal_db_error, Status, StatusCode};
 use ckb_logger::debug;
 use ckb_network::PeerIndex;
 use ckb_types::{packed, prelude::*};
@@ -34,15 +31,6 @@ impl<'a> BlockProcess<'a> {
         );
         let shared = self.synchronizer.shared();
         let state = shared.state();
-
-        let status = shared.active_chain().get_block_status(&block.hash());
-        if matches!(status, BlockStatus::UNKNOWN | BlockStatus::BLOCK_INVALID) {
-            return StatusCode::BlockIsInvalid.with_context(format!(
-                "BlockProcess refuse this block: block status is {:?}, block_hash: {}",
-                status,
-                block.hash(),
-            ));
-        }
 
         if state.new_block_received(&block) {
             if let Err(err) = self.synchronizer.process_new_block(block.clone()) {