+++ test/src/main.rs
@@ -362,7 +362,6 @@ fn canonicalize_path<P: AsRef<Path>>(path: P) -> PathBuf {
 
 fn all_specs() -> Vec<Box<dyn Spec>> {
     let mut specs: Vec<Box<dyn Spec>> = vec![
-        Box::new(BanPeerWhoSendBlockIDidNotRequest),
         Box::new(BlockSyncFromOne),
         Box::new(BlockSyncForks),
         Box::new(BlockSyncDuplicatedAndReconnect),